+++ include/nuclear-data-reader/ame_reaction2_position.hpp
@@ -8,7 +8,7 @@ namespace AME
 {
   struct Reaction2Position
   {
-    explicit Reaction2Position(const int _year)
+    explicit Reaction2Position(const uint16_t _year)
     {
       if (_year < 2020)
         {
